add BankID (Norway) to the list of apps blocking GrapheneOS
https://bankid.no/hjelp-bankid-app

My translation: 

"I have an Android phone, but can't activate biometrics. What could be wrong? 

Some Android phones lack technical support, or so-called 'key attestation', that is needed in order to activate biometrics. This is when Google confirms the security of the key exchange. For instance, if one modifies one's phone, it might occur that Google cannot accept it."